+++ src/main/kotlin/com/github/sleepypanda/feesh/constants/RareDrops.kt
@@ -15,6 +15,8 @@ enum class RareDropTypes(val displayName: String) {
     TIKI_MASK("Tiki Mask"),
     TITANOBOA_SHED("Titanoboa Shed"),
     SNAKE_EYES("Snake Eyes"),
+    OCTOPUS_TENDRIL("Octopus Tendril"),
+    TROUBLED_BUBBLE("Troubled Bubble"),
     SCUTTLER_SHELL("Scuttler Shell"),
     BURNT_TEXTS("Burnt Texts"),
     FLASH_1("Flash 1"),
@@ -22,6 +24,9 @@ enum class RareDropTypes(val displayName: String) {
     VIBRANT_CORAL("Vibrant Coral"),
     TRUE_ICE("True Ice"),
     PRINCES_CROWN_JEWEL("Prince's Crown Jewel"),
+    DISTANT_ECHO("Distant Echo"),
+    REINFORCED_NETTING("Reinforced Netting"),
+    WATER_HYACINTH("Water Hyacinth"),
 
     MEGALODON_LEGENDARY("Megalodon (Legendary)"),
     MEGALODON_EPIC("Megalodon (Epic)"),
@@ -123,6 +128,22 @@ class RareDrops {
                 defaultSoundFileName = Sounds.FEESH_MINECRAFT_CHALLENGE_COMPLETED
             ),
             RareDropInfo(
+                id = "OCTOPUS_TENDRIL",
+                itemName = RareDropTypes.OCTOPUS_TENDRIL.displayName,
+                rarityColorCode = LEGENDARY.code,
+                npcPrice = 1_000_000,
+                isExtremelyRare = false,
+                defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
+            ),
+            RareDropInfo(
+                id = "TROUBLED_BUBBLE",
+                itemName = RareDropTypes.TROUBLED_BUBBLE.displayName,
+                rarityColorCode = LEGENDARY.code,
+                npcPrice = 1_000_000,
+                isExtremelyRare = false,
+                defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
+            ),
+            RareDropInfo(
                 id = "SCUTTLER_SHELL",
                 itemName = RareDropTypes.SCUTTLER_SHELL.displayName,
                 rarityColorCode = MYTHIC.code,
@@ -169,6 +190,30 @@ class RareDrops {
                 itemName = RareDropTypes.TRUE_ICE.displayName,
                 rarityColorCode = LEGENDARY.code,
                 npcPrice = null,
+                isExtremelyRare = false,
+                defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
+            ),
+            RareDropInfo(
+                id = "WATER_HYACINTH",
+                itemName = RareDropTypes.WATER_HYACINTH.displayName,
+                rarityColorCode = LEGENDARY.code,
+                npcPrice = 1_000_000,
+                isExtremelyRare = false,
+                defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
+            ),
+            RareDropInfo(
+                id = "DISTANT_ECHO",
+                itemName = RareDropTypes.DISTANT_ECHO.displayName,
+                rarityColorCode = LEGENDARY.code,
+                npcPrice = 1_000_000,
+                isExtremelyRare = false,
+                defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
+            ),
+            RareDropInfo(
+                id = "REINFORCED_NETTING",
+                itemName = RareDropTypes.REINFORCED_NETTING.displayName,
+                rarityColorCode = LEGENDARY.code,
+                npcPrice = 1_000_000, // ?
                 isExtremelyRare = false,
                 defaultSoundFileName = Sounds.FEESH_OH_MY_GOD
             ),